add emulator.h that supports randomization
also it fixes a segfault in mem_t (addresses need to be masked).
for a memory whose depth is a non-power-of-2, say 2^n < d < 2^(n+1),
an attempt to read a value in the range [d, 2^(n+1)) will now return
a random number; a write will be ignored.  the check (and the call
to rand()) should be optimized away for power-of-2 depth RAMs.