improve Reg QoR; simplify genMuxes
This patch will break Verilog test harnesses that fail to delay input signal
transitions with respect to the rising clock edge.  It changes the Verilog
emitted for Regs to remove nodes and improve QoR with Synopsys DC.

Previously, Chisel code like the following

  val enable = Bool()
  val newData = Bits()
  val reg = Reg() { Bits() }
  when (enable) { reg := newData }

would produce Verilog like the following:

  assign t = enable ? newData : reg;
  always @(posedge clk)
    if (enable)
      reg <= t;

which is obvioulsy equivalent to the following:

  always @(posedge clk)
    if (enable)
      reg <= newData;

but Synopsys' tools add the extraneous mux.  This patch removes it.