Add debug interface
Experimental debug interface which adds a string -> dat_t or
string->mem_t mapping table. This could be used to peek or poke
arbitrary registers in the emulator.
Also adds an example interface (interactive_fragment.cpp) which allows
register reads/writes with console commands. This is currently very
brittle and clumsy because of limited string->dat_t support (but it does
work).