(Flo) Fix "eq" and "neq" width specification
The meaning of the width paramater in "eq" and "neq" nodes changed at
some point recently -- it used to be that the width was the width of
the input nodes, but now it's the output width.  As the output width
of all comparison operations is always 1 bit, this is pretty useless.

This patch changes back to the old behavior which is both more useful
and works with existing code.