Add "--dumpTestInput" as an argument
This will cause the stdin that's read by the tester to get dumped to a
file during every run.  The file will be named "Design.stdin", where
"Design" is the name of the top-level design that's being run (it's
the same as the rest of the generated files, like "Design.cpp" or
"Design.vcd").

I'm currently using this functionality in order to attempt to get
completely determintsic tests so I can run both the C++ backend and my
new LLVM backend and expect to get the same result.

It could be useful for running tests on embedded systems that can't
run Scala (like the Zedboard).