Add "--testerSeed <N>" command-line argument
The Scala tester uses a random number generator that's currently not
seeded.  This leads to non-determinstic tester behavior, which can be
a pain when you're trying to debug something.

This patch adds a "--testerSeed N" command-line option to Chisel.
This sets the RNG's seed to N.  Without this option the old behavior
is retained.