Add "--emitTempNodes" command-line option
As part of verifying the correctness of the LLVM backend I am
currently running a diff between the VCDs that Chisel's C++ backend
emits and those that my toolchain emits.  For some larger designs
there is so much computation that goes on inside temporary nodes it's
a bit hard to figure out exactly what's going on without them in the
VCD file.

This patch adds an "--emitTempNodes" command-line option, which emits
these temporary nodes into the VCD file.  Note that "--emitTempNodes"
implies "--debug" since it feels silly to emit these temporary nodes
without emiting the rest of the internal nodes.

This probably isn't that useful for users, but I can see how it'd be
nice for debugging Chisel internals.