Factor out global options/global state into Driver
Driver represents the "compiler driver."  Ultimately, there shouldn't be
global state at all; each Chisel run should produce a Driver object, and
each Module would have a reference to the Driver that owns it.  But that's
a more substantial change.

This commit touches a lot of code, but it's mostly just search-and-replace.