Fix tutorial Adder4 failure due to unmasked literal comparison.
Currently, literal values are allowed to overflow their width, so
comparing widths isn't sufficient to determine if we can use a literal
value directly during bit extraction. We need to also verify that the
literal value doesn't overflow the width. It may be more appropriate
to fix this elsewhere, but that requires a little more understanding
(and research) about "width" management.