Fix long RSH by adding dat_t::operator[](size_t)
I'm not entirely sure if this is the correct way to fix this, but it
seems to be the least intrusive way of going about it.  For some
reason the multi-word RSH code was emiting some operators that aren't
defined.  I couldn't figure out why, so I just went ahead and added
the operator in.