First "real" working version (dot and cpp backends).
Add input "plumbing" to each island that consumes it.
Create a dummy "global" island when we aren't partitioning to simplify
the logic. Write out the "top-level" non-island clock code.
Still some confusion over what/when we need to draw nodes when we draw
edges between them.
Just use "Mem" (not MemRead/MemWrite) as an island boundary - the read
needs to be associated with the code doing the reading, and the write
has to be associated with the code doing the writing.