Compatibility issues, multiple clock domains, top-level dot file module crossings (when partitioning), handling non-bits node input chains.
Output the monolithic cpp file (for compatibility testing).
Put shadow registers in the object to enable clock_hi to work with
partitioned island code where the clock init/def sequences are in
separate functions and there's no local variable to save the register
state in. Perhaps we should have clock-domain specific storage for the
registers where we can toggle active values by updating an index.
Fix bug in the multiple clock domain code that re-initialize the
islandClkCode every time we saw a new island, thereby effectively
forgetting all the previous island/clock domain specific code.
Deal with clock initHi and DefHi independently when partitioning - we
have to run all the initHi code before any of the defHi code.
We may end up with some top level nodes in separate islands when
partitioning. Deal with that case in the Dot backend.
When we're looking for a non-bits node from which to flood in
islandsFromNonBitsNode(), we may hit the end of the chain before
finding one. In this case, don't make the last bits node if it has no
consumers, otherwise, we'll emit that entire chain from the islands
and die with a NullPointer exception when we try to find the island
for that node.