Split large files.
Define LineLimitedFunction() class to help deal with large source
files.
Provide separate optimization flag for the file containing the clone()
method.
Add switches --lineLimitFunctions and --minimumLinesPerFile.
In an attempt to speed up c++ compilation by shrinking files
containing method implementations, we've increased the size of the
circuit object which seems to have a significantly adverse effect on
the clone() method. This requires more research.
Move the function prolog and epilog out of dumpVCDInit() and into
clients so it is amenable to line limiting.