Work around a left-shift code gen issue
I don't really understand what's going on here, but this appears to
work around a code gen issue that chisel-torture found a while ago.
It appears that for some reason wordMangle() doesn't do the right
thing when running in non-debug mode, so I just went ahead and worked
around the problem.

I'm pretty sure this is the wrong solution, but I have no idea what
the correct one is.

This fixes chisel bug #239