Fix issue 287 by adjusting reset assign logic
A function in the backend pass which assigns clocks and resets has now been renamed to better show that is forces an explicit reset. Also, this function was rewritten to prefer an explicit clock domain reset over the reset of the parent. Finally, some dead code in the delay clock/reset assignment pass has been pruned.