Skeleton fix for issue #265 by adjusting otherwise behavior
Uncomment line that causes top-level otherwise blocks to be used as a default for wires. I couldn't really see how this made much sense to begin with as it tended to make the otherwise statement semantics confusing.

This solves the issue for registers; however, if, in the issue example, register is replaced by io.out then a null pointer exception occurs. While the user in that case did write incorrect code (the wire lacks an assignment for one of the conditions), throwing a null pointer exception (during a backend traversal no less) is bad form.