Change Node.line from val to def to respect --lineNumbers option
The "line" function returns the line number that a Node was created on
using the stack trace. However, it checks the getLineNumbers option to
decide whether or not to actually return anything. Since this option is
not set until chiselMain is called, having this expression as a val will
result in "line" being evaluated at a point when "getLineNumbers" is
always false, so "line" will always be null. This commit fixes that
issue.