Ensure consistency between markedNodes and the IslandCollection result.
When building reference-chip with --partitionIslands, we were getting
a Java NoSuchElementException when we tried to move nodes between
islands, because some of the markedNodes were not in the
IslandCollection result.
Reorganize the code so those collections are more likely to be kept in
sync.
Verify that entries exist for the source and destination islands
before attempting to modify them - generate a ChiselError instead of
an exception.