Address issue with width-inference of right-shift
In the case where the width is computed via subtraction (for example, constant right-shifts), there is the possibility of encountering a 'negative width' (likely due to a user-code bug). The old behavior set the width to be uninferred and thus would eventually try again; however, this can rapidly lead to non-termination issues for width inference. The new behavior sets the width to 0. In both old and new behaviors, a warning is emitted.