Disallow comparisons between Nodes and non-Nodes
This isn't very Scala-y, but it prevents a common error that results from
relying on implicit conversions that fail to be invoked.  For example,
if you have an implicit conversion from Int to UInt, and write something
like (UInt(1) != 0), the implicit conversion might not be invoked and
the expression will evaluate, unexpectedly, to Scala-Boolean(true), rather
than Chisel-Bool(false).

Now, the above will trigger a Chisel error if the implicit conversion is
not invoked.