In CPP, better detection of signals to randomize
The old version used the archaic isInObject function to help decide what wires to randomize in the CPP backend.
This caused completely inappropriate behavior when debug modes were turned on.
The new version uses a saner check. Now, all wires with no inputs (as before) and also that are not top-level inputs are randomized.