make ChiselErrors private so clients are forced to use Chisel.error() - #393
Several clients were adding errors to ChiselError.ChiselErrors
directly which unfortunately doesn't update ChiselError.hasError which
is used by ChiselError.checkpoint() to determine if it should abort or
not. We could have changed hasErrors to basically do a
ChiselErrors.isEmpty, but the ArrauBuffer contains warnings as well so
currently, hasErrors != isEmpty.