Generate gtkwave compatible --vcdMem signals.
Roms and Rams are currently named like this in th VCD file:

reg[0]
reg[1]
etc.

Gtkwave is an open source and AFAIK popular waveform viewer, but it
has a bug where it isn't able parse signals with the '['
character. This commit changes the naming to be like:

reg(0)
reg(1)
etc.

which it is able to parse.