Implement rsha_n correctly
I'm actually not sure this is correct, but the other version is
obviously wrong: it regularly shifts by an extremely large value (as the
result of unsigned underflow).  This actually manifests in rocket-chip,
Howie found and isolated a test case.

I found this using "-fsanitize=undefined" (kind of by accident, since as
far as I can tell this was just wrong), which seems like something that
should be run on emulator.h.  It's unfortunately a dynamic analysis, but
we should probably run it on rocket-chip.