Return Seq, not Vec, from PriorityEncoderOH
The return type actually used to be Seq, but it was inadvertently changed
to Vec at some point.

It doesn't really make sense to create a one-hot result then follow it
with a decoder.  It's more performant to use a PriorityEncoder followed
by a comparator.  Discourage the former by returning Seq, not Vec.